==> SageMaker/Training-Inference/0. Data Prep.ipynb <==
### Imports 

from sklearn.preprocessing import LabelEncoder, MinMaxScaler
from sklearn.model_selection import train_test_split
import pandas as pd
import random

### Load & Prep Data

The <a href='https://homepages.inf.ed.ac.uk/imurray2/teaching/oranges_and_lemons/'>Fruits Dataset</a> was originally created by Dr. Iain Murray from University of Edinburgh and extended more recently by the University of Michigan. It is a simple multi-class dataset with 4 columns (features) and 4 classes (fruits). The 4 classes are apple, orange, mandarin and lemon. The four features are mass, width, height and color score of the fruit.

The color score feature maps to a color and its intensity in the color spectrum (0 - 1) scale. <br><br>
<table align="left" style="width:50%">
    <tr>
        <th>Color</th>
        <th>Range</th>
    </tr>
    <tr>
        <td>Red</td>
        <td>0.85 - 1.00</td>
    </tr>
    <tr>
        <td>Orange</td>
        <td>0.75 - 0.85</td>
    </tr>
    <tr>
        <td>Yellow</td>
        <td>0.65 - 0.75</td>
    </tr>
    <tr>
        <td>Green</td>
        <td>0.45 - 0.65</td>
    </tr>
</table>

df = pd.read_csv('./DATA/fruits.csv', names=['class', 'mass', 'width', 'height', 'color_score'])

df.head()

df.shape

df['class'].unique().tolist()

X = df[['mass', 'width', 'height', 'color_score']]
y = df['class']

#### Encode the classes into numerical values using Sklearn's LabelEncoder

label_encoder = LabelEncoder()
label_encoder.fit(['apple', 'orange', 'mandarin', 'lemon'])
y = label_encoder.transform(y)

y

#### Split X, y into train and test sets

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=123)

X_train.shape

X_test.shape

#### Scale feature columns using Sklearn's MinMaxScaler

scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

X_train[0]

y_train[0]

#### Combine Scaled X & y into Train and Test DataFrames 

X_train = pd.DataFrame(X_train, columns=['mass', 'width', 'height', 'color_score'])
y_train = pd.DataFrame(y_train, columns=['class'])
train_df = pd.concat([y_train, X_train], axis=1)
train_df.head()

X_test = pd.DataFrame(X_test, columns=['mass', 'width', 'height', 'color_score'])
y_test = pd.DataFrame(y_test, columns=['class'])
test_df = pd.concat([y_test, X_test], axis=1)
test_df.head()

#### Create a DataFrame for Batch Inference without the Class column

batch_test_df = test_df.drop(['class'], axis=1)
batch_test_df.head()

#### Write Train & Test Sets to Local Directory

train_df.to_csv('./DATA/train/train.csv', header=False, index=False)
test_df.to_csv('./DATA/test/test.csv', header=False, index=False)
batch_test_df.to_csv('./DATA/batch_test/batch_test.csv', header=False, index=False)

<b>Write train_df without class label and with header for Model Monitoring Baselining</b>

train_df.to_csv('./DATA/train/train_with_header.csv', header=True, index=False)

### Let us simulate some artificial data for Model Monitor for our Data Shift experiments later

def get_random_val():
    val = random.uniform(0, 1)
    return round(val, 6)

def generate_row():
    row = []
    for _ in range(4):
        row.append(get_random_val())
    return row  

def generate_dataset():
    rows = []
    for _ in range(20):
        rows.append(generate_row())
    return rows 

rows = generate_dataset() 

df = pd.DataFrame(rows, columns=['mass', 'width', 'height', 'color_score'])

df.head()

df.to_csv('./DATA/test/model_monitor_test.csv', header=False, index=False)
