==> SageMaker/Training-Inference/1. Sklearn Local.ipynb <==
### Imports 

from sklearn.neighbors import KNeighborsClassifier
import pandas as pd

### Load Prepped Data

train_df = pd.read_csv('./DATA/train/train.csv', names=['class', 'mass', 'width', 'height', 'color_score'])
train_df.head()

X_train = train_df[['mass', 'width', 'height', 'color_score']]
y_train = train_df['class']

### Train Model (K Nearest Neighbor)

knn = KNeighborsClassifier()

knn.fit(X_train, y_train)

### Evaluate Model on Test Set

test_df = pd.read_csv('./DATA/test/test.csv', names=['class', 'mass', 'width', 'height', 'color_score'])

X_test = train_df[['mass', 'width', 'height', 'color_score']]
y_test = train_df['class']

acc = knn.score(X_test, y_test)
print('Accuracy = {:.2f}%'.format(acc * 100))
