==> SageMaker/Training-Inference/2. Built-in Algorithm XGBoost.ipynb <==
### 1. Imports 

# ML Imports 
import pandas as pd
import numpy as np
import json
import os

# AWS Imports 
from sagemaker.amazon.amazon_estimator import get_image_uri
from sagemaker.predictor import csv_serializer
from sagemaker import get_execution_role
import sagemaker
import boto3

### 2. Essentials 

# Get the Execution Role for this Notebook and AWS Session to create Clients
sagemaker_execution_role = get_execution_role()
print('Role = {}'.format(sagemaker_execution_role))
session = boto3.Session()

# Clients and Resources
s3 = session.resource('s3')
sagemaker_session = sagemaker.Session()

BUCKET = 'sagemaker-demo-892313895307' # USE YOUR ACCOUNT ID OR INITIALS AS SUFFIX
PREFIX = 'clf'

##### Upload Train & Test Sets to S3 and Create Pointers to Data

s3.create_bucket(Bucket=BUCKET)

s3.Bucket(BUCKET).Object(os.path.join(PREFIX, 'train/train.csv')).upload_file('./DATA/train/train.csv')
s3.Bucket(BUCKET).Object(os.path.join(PREFIX, 'test/test.csv')).upload_file('./DATA/test/test.csv')
s3.Bucket(BUCKET).Object(os.path.join(PREFIX, 'batch_test/batch_test.csv')).upload_file('./DATA/batch_test/batch_test.csv')

##### Create Pointers to the uploaded files 

train_set_location = 's3://{}/{}/train/'.format(BUCKET, PREFIX)
test_set_location = 's3://{}/{}/test/'.format(BUCKET, PREFIX)

print(train_set_location)
print(test_set_location)

train_set_pointer = sagemaker.s3_input(s3_data=train_set_location, content_type='csv')
test_set_pointer = sagemaker.s3_input(s3_data=test_set_location, content_type='csv')

print(json.dumps(train_set_pointer.__dict__, indent=2))

### 3. Train a Model using SageMaker + Builtin XgBoost Algorithm

container_uri = get_image_uri(session.region_name, 'xgboost', '1.0-1')

xgb = sagemaker.estimator.Estimator(image_name=container_uri,
                                    role=sagemaker_execution_role, 
                                    train_instance_count=2, 
                                    train_instance_type='ml.m5.large',
                                    output_path='s3://{}/{}/model-artifacts'.format(BUCKET, PREFIX),
                                    sagemaker_session=sagemaker_session,
                                    base_job_name='classifier')

xgb.set_hyperparameters(objective='multi:softmax',
                        num_class=4,
                        num_round=100)

xgb.fit({'train': train_set_pointer, 'validation': test_set_pointer})

### 4. Host the Trained Model as a SageMaker Endpoint (using Estimator object)

xgb_predictor = xgb.deploy(initial_instance_count=2,
                           instance_type='ml.m5.large')

### 5. Real Time Inference from the Endpoint

xgb_predictor.content_type = 'text/csv'
xgb_predictor.serializer = csv_serializer

test_df = pd.read_csv('./DATA/test/test.csv', names=['class', 'mass', 'width', 'height', 'color_score'])

test_df.head()

X = test_df.sample(1)
X

X = X.values[0]
X[1:]

payload = X[1:]

%%time

predicted_class = xgb_predictor.predict(payload).decode('utf-8')
predicted_class

### 5. Evaluate Hosted Model for Accuracy

predictions = []
expected = []
correct = 0
for row in test_df.values:
    expected_class = row[0]
    payload = row[1:]
    predicted_class = xgb_predictor.predict(payload).decode('utf-8')
    if predicted_class == str(expected_class):
        correct += 1
    predictions.append(predicted_class)
    expected.append(expected_class)

print('Accuracy = {:.2f}%'.format(correct/len(predictions) * 100))

#### Confusion Matrix

expected = pd.Series(np.array(expected))
predictions = pd.Series(np.array(predictions))
pd.crosstab(expected, predictions, rownames=['Actual'], colnames=['Predicted'], margins=True)
