==> SageMaker/Training-Inference/3. Endpoint Configuration.ipynb <==
### Decouple Training & Hosting

<b> Use Scenario</b><br>
When you want to create an endpoint using Model artifacts trained previously

<b>Note:</b> Let's use the XgBoost Classifier model artifacts we trained previously to create an endpoint here.

### Imports 

from time import gmtime, strftime
import sagemaker

### 1. Essentials

current_timestamp = strftime("%Y-%m-%d-%H-%M-%S", gmtime())

TRAINING_JOB_NAME = 'fruits-clf-xgboost-2020-11-04-15-45-32-296' # COPY THIS FROM THE CONSOLE
MODEL_NAME = f'fruits-clf-xgboost-model-{current_timestamp}'
ENDPOINT_CONFIG_NAME = f'fruits-clf-xgboost-endpoint-config-{current_timestamp}'
ENDPOINT_NAME = f'fruits-clf-xgboost-endpoint-{current_timestamp}'

sagemaker_session = sagemaker.session.Session()

### 2. Create a pointer to your model artifacts using the previously trained job name

sagemaker_session.create_model_from_job(training_job_name=TRAINING_JOB_NAME, 
                                        name=MODEL_NAME)

### 3. Create an Endpoint Config

sagemaker_session.create_endpoint_config(ENDPOINT_CONFIG_NAME, 
                                         model_name=MODEL_NAME, 
                                         initial_instance_count=1, 
                                         instance_type='ml.r5.large')

### 4. Create a new Endpoint 

sagemaker_session.create_endpoint(ENDPOINT_NAME, ENDPOINT_CONFIG_NAME)
