==> SageMaker/Training-Inference/4. Batch Transform.ipynb <==
### Imports 

from sagemaker.amazon.amazon_estimator import get_image_uri
from sagemaker import get_execution_role
from time import gmtime, strftime
import pandas as pd
import sagemaker
import boto3
import time

### 1. Essentials

BUCKET = 'sagemaker-demo-892313895307'
PREFIX = 'fruits-clf'
REGION = 'us-east-1'

batch_input = f's3://{BUCKET}/{PREFIX}/batch_test/'
batch_input

batch_output = f's3://{BUCKET}/{PREFIX}/batch_test_out/'
batch_output

current_timestamp = strftime("%Y-%m-%d-%H-%M-%S", gmtime())

TRAINING_JOB_NAME = 'fruits-clf-xgboost-2020-11-04-15-45-32-296'  # COPY THIS FROM THE CONSOLE
MODEL_NAME = f'fruits-clf-xgboost-model-{current_timestamp}'
BATCH_JOB_NAME = f'fruits-clf-xgboost-batch-job-{current_timestamp}'

sagemaker_execution_role = get_execution_role()
sagemaker_session = sagemaker.session.Session()
sagemaker_client = boto3.client('sagemaker', region_name=REGION)
s3_client = boto3.client('s3')

container_uri = get_image_uri(REGION, 'xgboost', '1.0-1')

### 2. Create a Model object using previously run training job name

info = sagemaker_client.describe_training_job(TrainingJobName=TRAINING_JOB_NAME)
info

model_artifact_url = info['ModelArtifacts']['S3ModelArtifacts']
model_artifact_url

primary_container = {
    'Image': container_uri,
    'ModelDataUrl': model_artifact_url
  }

response = sagemaker_client.create_model(
    ModelName=MODEL_NAME,
    ExecutionRoleArn=sagemaker_execution_role,
    PrimaryContainer=primary_container)

response

### 3. Create a Batch Transformer for Inference

request = {
    "TransformJobName": BATCH_JOB_NAME,
    "ModelName": MODEL_NAME,
    "BatchStrategy": "MultiRecord",
    "TransformOutput": {
        "S3OutputPath": batch_output
    },
    "TransformInput": {
        "DataSource": {
            "S3DataSource": {
                "S3DataType": "S3Prefix",
                "S3Uri": batch_input 
            }
        },
        "ContentType": "text/csv",
        "SplitType": "Line",
        "CompressionType": "None"
    },
    "TransformResources": {
            "InstanceType": "ml.m5.xlarge",
            "InstanceCount": 1
    }
}

response = sagemaker_client.create_transform_job(**request)
response

while(True):
    response = sagemaker_client.describe_transform_job(TransformJobName=BATCH_JOB_NAME)
    status = response['TransformJobStatus']
    if  status == 'Completed':
        print("Transform job ended with status: {}".format(status))
        break
    if status == 'Failed':
        message = response['FailureReason']
        print('Transform failed with the following error: {}'.format(message))
        raise Exception('Transform job failed') 
    print("Transform job is still in status: {}".format(status))    
    time.sleep(30) 

### 4. Evaluate Output

key = f'{PREFIX}/batch-inference/batch_test.csv.out'

obj = s3_client.get_object(Bucket=BUCKET, Key=key)
results_df = pd.read_csv(obj['Body'], names=['Predictions'])

results_df
